move decimal_encode -> utils/decimal_encode edit imports
